+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NonPublicExtensions/Sources/NonPublicDacExtension.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using PX.Data;
 
-namespace PX.Objects.HackathonDemo.Extensions.NonPublic
+namespace Acuminator.Tests.Tests.StaticAnalysis.DacUiAttributes.Sources
 { 
 	internal sealed class SOOrderExt : PXCacheExtension<SOOrder>	//Non public DAC extensions are not supported
 	{
@@ -44,5 +44,49 @@ namespace PX.Objects.HackathonDemo.Extensions.NonPublic
 				#endregion
 			}
 		}
+	}
+
+
+
+	[PXHidden]
+	public class SOOrder : IBqlTable
+	{
+		#region OrderType
+		public abstract class orderType : IBqlField { }
+		[PXDBString(IsKey = true, InputMask = "")]
+		[PXDefault]
+		[PXUIField(DisplayName = "Order Type")]
+		public string OrderType { get; set; }
+		#endregion
+
+		#region OrderNbr
+		public abstract class orderNbr : IBqlField { }
+		[PXDBString(IsKey = true, InputMask = "")]
+		[PXDefault]
+		[PXUIField(DisplayName = "Order Nbr.")]
+		public string OrderNbr { get; set; }
+		#endregion
+
+		#region Status
+		public abstract class status : IBqlField { }
+
+		[PXStringList(new[] { "N", "O" }, new[] { "New", "Open" })]
+		[PXUIField(DisplayName = "Status")]
+		[PXString]
+		public string Status { get; set; }
+		#endregion
+
+		#region tstamp
+		public abstract class Tstamp : IBqlField
+		{
+		}
+
+		[PXDBTimestamp]
+		public virtual byte[] tstamp
+		{
+			get;
+			set;
+		}
+		#endregion
 	}
 }
