--- src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NonPublicExtensions/Sources/SOOrder.cs
@@ -1,51 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using PX.Data;
-
-namespace PX.Objects.HackathonDemo.Extensions.NonPublic
-{
-	[PXHidden]
-	public class SOOrder : IBqlTable
-	{
-        #region OrderType
-        public abstract class orderType : IBqlField { }
-		[PXDBString(IsKey = true, InputMask = "")]
-		[PXDefault]
-		[PXUIField(DisplayName = "Order Type")]
-		public string OrderType { get; set; }
-		#endregion
-
-		#region OrderNbr
-		public abstract class orderNbr : IBqlField { }
-		[PXDBString(IsKey = true, InputMask = "")]
-		[PXDefault]
-		[PXUIField(DisplayName = "Order Nbr.")]
-		public string OrderNbr { get; set; }
-		#endregion
-
-		#region Status
-		public abstract class status : IBqlField { }
-
-		[PXStringList(new[] { "N", "O" }, new[] { "New", "Open" })]
-		[PXUIField(DisplayName = "Status")]
-		[PXString]
-		public string Status { get; set; }
-		#endregion
-
-        #region tstamp
-        public abstract class Tstamp : IBqlField
-		{
-		}
-
-		[PXDBTimestamp]
-		public virtual byte[] tstamp
-		{
-			get;
-			set;
-		}
-		#endregion
-	}
-}
